refactor(ui): address experiment metrics review feedback
