+++ app/src/pages/dataset/metrics/ExperimentCostChart.tsx
@@ -24,13 +24,11 @@ import {
   floatShortFormatter,
 } from "@phoenix/utils/numberFormatUtils";
 
-import {
-  experimentXAxisProps,
-  useExperimentMetricsData,
-} from "./ExperimentMetrics";
 import { makeExperimentMetricsTooltipContent } from "./ExperimentMetricsTooltipContent";
+import { experimentXAxisProps } from "./experimentXAxisProps";
 import type { ExperimentMetricViewProps } from "./types";
 import { EXPERIMENT_METRICS_CHART_SYNC_ID } from "./types";
+import { useExperimentMetricsData } from "./useExperimentMetricsData";
 
 const TooltipContent = makeExperimentMetricsTooltipContent(costFormatter);
